+++ src/main/kotlin/net/alexanderkahn/longball/service/model/LineupPlayer.kt
@@ -1,4 +1,4 @@
 package net.alexanderkahn.longball.service.model
 
-data class LineupPlayer(val player: Long, val battingOrder: Short, val fieldPosition: FieldPosition)
+data class LineupPlayer(val player: Long, val battingOrder: Int, val fieldPosition: FieldPosition)
